MogatoRoboto game
Mogato Roboto is a 3D game written in C# using the Mogre engine. You play as an alien character in 3rd person. The objective is to collect as many gems as possible, whilst avoiding enemy robots. weapons such as a grenade and cannon can be acquired to destroy enemies. 

This project was an extension of the 3D Modeling and Rendering components.